+++ crates/hyperqueue/src/server/client/submit.rs
@@ -99,7 +99,7 @@ pub async fn handle_submit(
         .unwrap()
     {
         ToGatewayMessage::NewTasksResponse(_) => { /* Ok */ }
-        _ => panic!("Invalid response"),
+        r => panic!("Invalid response: {:?}", r),
     };
 
     ToClientMessage::SubmitResponse(SubmitResponse { job: job_detail })