+++ crates/pyhq/python/hyperqueue/ffi/protocol.py
@@ -1,12 +1,24 @@
 import dataclasses
-from typing import Dict, List, Optional, Sequence
+from typing import Dict, List, Optional, Sequence, Union
 
 
 @dataclasses.dataclass()
 class ResourceRequest:
     n_nodes: int = 0
-    cpus: str = "1"
-    generic: Dict[str, int] = dataclasses.field(default_factory=dict)
+    resources: Dict[str, Union[int, str]] = dataclasses.field(default_factory=dict)
+
+    def __init__(
+        self,
+        *,
+        n_nodes=0,
+        cpus: Union[int, str] = 1,
+        resources: Optional[Dict[str, Union[int, str]]] = None
+    ):
+        self.n_nodes = n_nodes
+        if resources is None:
+            resources = {}
+        resources["cpus"] = cpus
+        self.resources = resources
 
 
 @dataclasses.dataclass()